# backend/docs/swagger.yaml
basePath: /
definitions:
...
  github_com_zcl0621_compx576-smart-dairy-system_dto_report.ListResponse:
    properties:
      list:
        items:
          $ref: '#/definitions/github_com_zcl0621_compx576-smart-dairy-system_dto_report.ReportItem'
        type: array
      page:
        type: integer
      total:
        type: integer
      total_pages:
        type: integer
    type: object
  github_com_zcl0621_compx576-smart-dairy-system_dto_report.ReportItem:
    properties:
      cow_id:
        type: string
      cow_name:
        type: string
      created_at:
        type: string
      details:
        $ref: '#/definitions/model.ReportDetails'
      id:
        type: string
      period_end:
        type: string
      period_start:
        type: string
      score:
        type: number
      summary:
        type: string
      updated_at:
        type: string
    type: object
  github_com_zcl0621_compx576-smart-dairy-system_dto_user.CreateRequest:
    properties:
      email:
        type: string
      password:
        type: string
      username:
        type: string
    required:
    - email
    - password
    - username
    type: object
  github_com_zcl0621_compx576-smart-dairy-system_dto_user.DeleteRequest:
    properties:
      id:
        type: string
    required:
    - id
    type: object
  github_com_zcl0621_compx576-smart-dairy-system_dto_user.InfoResponse:
    properties:
      created_at:
        type: string
      email:
        type: string
      id:
        type: string
      updated_at:
        type: string
      username:
        type: string
    type: object
  github_com_zcl0621_compx576-smart-dairy-system_dto_user.ListItem:
    properties:
      created_at:
        type: string
      email:
        type: string
      id:
        type: string
      updated_at:
        type: string
      username:
        type: string
    type: object
  github_com_zcl0621_compx576-smart-dairy-system_dto_user.ListResponse:
    properties:
      list:
        items:
          $ref: '#/definitions/github_com_zcl0621_compx576-smart-dairy-system_dto_user.ListItem'
        type: array
      page:
        type: integer
      total:
        type: integer
      total_pages:
        type: integer
    type: object
  github_com_zcl0621_compx576-smart-dairy-system_dto_user.UpdatePasswordRequest:
    properties:
      id:
        type: string
      password:
        type: string
    required:
    - id
    - password
    type: object
  github_com_zcl0621_compx576-smart-dairy-system_dto_user.UpdateRequest:
    properties:
      email:
        type: string
      id:
        type: string
      username:
        type: string
    required:
    - email
    - id
    - username
    type: object
  model.AlertSeverity:
    enum:
    - warning
    - critical
    - offline
    type: string
    x-enum-varnames:
    - AlertSeverityWarning
    - AlertSeverityCritical
    - AlertSeverityOffline
...
  model.MetricType:
    enum:
    - temperature
    - heart_rate
    - blood_oxygen
    - latitude
    - longitude
    - milk_amount
    - milking_duration
    - weight
    - device
    type: string
    x-enum-varnames:
    - MetricTypeTemperature
    - MetricTypeHeartRate
    - MetricTypeBloodOxygen
    - MetricTypeLatitude
    - MetricTypeLongitude
    - MetricTypeMilkAmount
    - MetricTypeMilkingDuration
    - MetricTypeWeight
    - MetricTypeDevice
  model.MetricUnit:
    enum:
    - celsius
    - bpm
    - percent
    - degrees
    - liters
    - seconds
    - kg
    type: string
    x-enum-varnames:
    - MetricUnitCelsius
    - MetricUnitBPM
    - MetricUnitPercent
    - MetricUnitDegrees
    - MetricUnitLiters
    - MetricUnitSeconds
    - MetricUnitKG
  model.ReportAlert:
    properties:
      level:
        $ref: '#/definitions/model.AlertSeverity'
      message:
        type: string
    type: object
  model.ReportDetails:
    properties:
      alerts:
        items:
          $ref: '#/definitions/model.ReportAlert'
        type: array
      metrics:
        items:
          $ref: '#/definitions/model.ReportMetric'
        type: array
      note:
        type: string
    type: object
  model.ReportMetric:
    properties:
      key:
        $ref: '#/definitions/model.MetricType'
      label:
        type: string
      status:
        $ref: '#/definitions/model.ReportMetricStatus'
      text:
        type: string
      unit:
        $ref: '#/definitions/model.MetricUnit'
      value:
        type: number
    type: object
  model.ReportMetricStatus:
    enum:
    - normal
    - warning
    - critical
    - offline
    type: string
    x-enum-varnames:
    - ReportMetricStatusNormal
    - ReportMetricStatusWarning
    - ReportMetricStatusCritical
    - ReportMetricStatusOffline
  runtime_web_server_handler.ErrorResponse:
    properties:
      error:
        type: string
    type: object
  runtime_web_server_handler.OKResponse:
    properties:
      ok:
        type: boolean
    type: object
info:
  contact: {}
  description: api for smart dairy cows
  title: smart dairy api
  version: "1.0"
paths:
...
  /api/report/latest:
    get:
      description: get latest report for one cow
      parameters:
      - description: Cow ID
        in: query
        name: cow_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_zcl0621_compx576-smart-dairy-system_dto_report.ReportItem'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/runtime_web_server_handler.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/runtime_web_server_handler.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/runtime_web_server_handler.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/runtime_web_server_handler.ErrorResponse'
      security:
      - BearerAuth: []
      summary: get latest report
      tags:
      - Report
  /api/report/list:
    get:
      description: get paged report list
      parameters:
      - default: 1
        description: page numb
        in: query
        name: page
        type: integer
      - default: 20
        description: page size
        in: query
        name: page_size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_zcl0621_compx576-smart-dairy-system_dto_report.ListResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/runtime_web_server_handler.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/runtime_web_server_handler.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/runtime_web_server_handler.ErrorResponse'
      security:
      - BearerAuth: []
      summary: list reports
      tags:
      - Report
  /api/user/create:
    post:
      consumes:
      - application/json
      description: create one user
      parameters:
      - description: create user body
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_zcl0621_compx576-smart-dairy-system_dto_user.CreateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/runtime_web_server_handler.OKResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/runtime_web_server_handler.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/runtime_web_server_handler.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/runtime_web_server_handler.ErrorResponse'
      security:
      - BearerAuth: []
      summary: create user
      tags:
      - User
  /api/user/delete:
    post:
      consumes:
      - application/json
      description: delete one user
      parameters:
      - description: delete user body
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_zcl0621_compx576-smart-dairy-system_dto_user.DeleteRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/runtime_web_server_handler.OKResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/runtime_web_server_handler.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/runtime_web_server_handler.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/runtime_web_server_handler.ErrorResponse'
      security:
      - BearerAuth: []
      summary: delete user
      tags:
      - User
  /api/user/info:
    get:
      description: get one user by id
      parameters:
      - description: User ID
        in: query
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_zcl0621_compx576-smart-dairy-system_dto_user.InfoResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/runtime_web_server_handler.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/runtime_web_server_handler.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/runtime_web_server_handler.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/runtime_web_server_handler.ErrorResponse'
      security:
      - BearerAuth: []
      summary: get user info
      tags:
      - User
  /api/user/list:
    get:
      description: get paged user list
      parameters:
      - default: 1
        description: page numb
        in: query
        name: page
        type: integer
      - default: 20
        description: page size
        in: query
        name: page_size
        type: integer
      - description: user name
        in: query
        name: name
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_zcl0621_compx576-smart-dairy-system_dto_user.ListResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/runtime_web_server_handler.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/runtime_web_server_handler.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/runtime_web_server_handler.ErrorResponse'
      security:
      - BearerAuth: []
      summary: list users
      tags:
      - User
  /api/user/update:
    post:
      consumes:
      - application/json
      description: update one user
      parameters:
      - description: update user body
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_zcl0621_compx576-smart-dairy-system_dto_user.UpdateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/runtime_web_server_handler.OKResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/runtime_web_server_handler.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/runtime_web_server_handler.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/runtime_web_server_handler.ErrorResponse'
      security:
      - BearerAuth: []
      summary: update user
      tags:
      - User
  /api/user/update_password:
    post:
      consumes:
      - application/json
      description: update password for one user
      parameters:
      - description: update password body
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_zcl0621_compx576-smart-dairy-system_dto_user.UpdatePasswordRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/runtime_web_server_handler.OKResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/runtime_web_server_handler.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/runtime_web_server_handler.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/runtime_web_server_handler.ErrorResponse'
      security:
      - BearerAuth: []
      summary: update user password
      tags:
      - User
...
schemes:
- http
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"